+++ apps/kokomen-native/src/hooks/uselogin.ts
@@ -1,0 +1,8 @@
+export default function useLogin() {
+  function appleLogin() {
+    console.log("appleLogin");
+  }
+  return {
+    appleLogin,
+  };
+}
