+++ apps/kokomen-native/src/screens/interviews/interviewMain.tsx
@@ -5,9 +5,8 @@ import {
   SafeAreaView,
   View,
 } from "react-native";
-import WebView, { WebViewMessageEvent } from "react-native-webview";
-import useSpeechRecognition from "@/hooks/useSpeechRecognition";
-import { ExpoSpeechRecognitionModule } from "expo-speech-recognition";
+import WebView from "react-native-webview";
+import useWebviewEvents from "@/hooks/useWebviewEvents";
 
 export default function InterviewMainScreen() {
   const webviewRef = useRef<WebView>(null);
@@ -16,84 +15,7 @@ export default function InterviewMainScreen() {
       true;
     `;
 
-  const { handleStart, handleStop, isListening } = useSpeechRecognition({
-    onResult: (transcript) => {
-      const speechRecognitionResult = JSON.stringify({
-        type: "speechRecognitionResult",
-        data: transcript,
-      });
-      webviewRef.current?.postMessage(speechRecognitionResult);
-    },
-    onStart: () => {
-      webviewRef.current?.postMessage(
-        JSON.stringify({
-          type: "startListening",
-        }),
-      );
-    },
-    onEnd: () => {
-      webviewRef.current?.postMessage(
-        JSON.stringify({
-          type: "stopListening",
-        }),
-      );
-    },
-  });
-  const checkSpeechRecognitionSupported = () => {
-    ExpoSpeechRecognitionModule.requestPermissionsAsync()
-      .then((result) => {
-        if (result.status === "granted") {
-          webviewRef.current?.postMessage(
-            JSON.stringify({
-              type: "checkSpeechRecognitionSupported",
-              data: true,
-            }),
-          );
-        } else {
-          webviewRef.current?.postMessage(
-            JSON.stringify({
-              type: "checkSpeechRecognitionSupported",
-              data: false,
-            }),
-          );
-        }
-      })
-      .catch(() => {
-        webviewRef.current?.postMessage(
-          JSON.stringify({
-            type: "checkSpeechRecognitionSupported",
-            data: false,
-          }),
-        );
-      });
-  };
-
-  const pageChange = () => {
-    if (isListening) {
-      handleStop();
-    }
-  };
-  const handleMessage = (event: WebViewMessageEvent) => {
-    try {
-      const data = JSON.parse(event.nativeEvent.data);
-      switch (data.type) {
-        case "startListening":
-          handleStart();
-          break;
-        case "stopListening":
-          handleStop();
-          break;
-        case "checkSpeechRecognitionSupported":
-          checkSpeechRecognitionSupported();
-          break;
-        case "pageChange":
-          pageChange();
-          break;
-      }
-    } catch (error) {
-      console.error("error while parsing message", error);
-    }
-  };
+  const { handleMessage } = useWebviewEvents(webviewRef);
 
   return (
     <View style={{ flex: 1, backgroundColor: "#fff" }}>
